+++ src/com/andrewjrowell/fly/entities/RottenManager.java
@@ -36,7 +36,7 @@ public class RottenManager {
 	 * @param deltaTime time since last update
 	 * @param pace speed modifier
 	 */
-	public void update(float deltaTime, int pace){
+	public void update(float deltaTime, float pace){
 		rottencounter += deltaTime * pace / 48;
 		// Divide by 48 so at normal pace, counter
 		// represents time in seconds
